Set pytest marks. Pushing to try utub_ui tests
